reverse topological order for scheduling
In the MapPolicy with finite maxconcurrency set,
the scheduler will now try to fill execution stages
from the beginning to the end, except vice versa.

As a result, there is a greater priority to finish
late tasks early. This hopefully reduces peak memory consumption
and, if the workflow contains CachedBlueprints, it avoids
scheduling the caching at the very end, so that intermediate
results are saved before the job is done.

It is not quite clear if this will have negative impact
on other workloads, but this seems to be not the case for now